Refactored APilotEntry
renamed APilotEntry struct and moved it into seperate file. Renamed Decl.h to decl.h and adjusted includes
