+++ src/experimental/aentry.cpp
@@ -29,16 +29,4 @@ const TableData& AEntry::getData()
     return tableData;
 }
 
-APilotEntry::APilotEntry()
-    : AEntry::AEntry(DEFAULT_PILOT_POSITION)
-{}
-
-APilotEntry::APilotEntry(int row_id)
-    : AEntry::AEntry(DataPosition("pilots", row_id))
-{}
-
-APilotEntry::APilotEntry(TableData table_data)
-    : AEntry::AEntry(DEFAULT_PILOT_POSITION, table_data)
-{}
-
 }  // namespace experimental
